fix(AddRole): organising the file structure
