+++ src/pages/privileges/outlets/roles/add-role/config/initialValues.ts
@@ -1,0 +1,20 @@
+import { IFormAddRole } from "@pages/privileges/outlets/roles/types";
+
+export const initialValuesAddRol: IFormAddRole = {
+  generalInformation: {
+    isValid: false,
+    values: {
+      roleName: "",
+      description: "",
+      aplication: "",
+    },
+  },
+  ancillaryAccounts: {
+    isValid: false,
+    values: {
+      officialSector: "",
+      commercialSector: "",
+      solidaritySector: "",
+    },
+  },
+};
