+++ src/pages/privileges/outlets/roles/add-role/index.tsx
@@ -1,120 +1,13 @@
-import { FormikProps } from "formik";
 import { useRef, useState } from "react";
+import { FormikProps } from "formik";
 
-import { IFormAddRoleRef } from "../types";
+import { IFormAddRole, IFormAddRoleRef } from "../types";
+import { addRoleStepsRules } from "../utils";
 import { stepsAddRol } from "./config/addRol.config";
 import { IGeneralInformationForm } from "./forms/GeneralInformationForm";
 import { IAncillaryAccountsForm } from "./forms/AncillaryAccounts";
 import { AddRolUI } from "./interface";
-
-const initialValuesAddRol: IFormAddRole = {
-  generalInformation: {
-    isValid: false,
-    values: {
-      roleName: "",
-      description: "",
-      aplication: "",
-    },
-  },
-  ancillaryAccounts: {
-    isValid: false,
-    values: {
-      officialSector: "",
-      commercialSector: "",
-      solidaritySector: "",
-    },
-  },
-};
-
-const creditDestinationStepsRules = (
-  currentStep: number,
-  addLinixRole: IFormAddRole,
-  formReferences: IFormAddRoleRef,
-  isCurrentFormValid: boolean
-) => {
-  let newAddLinixRole = { ...addLinixRole };
-
-  switch (currentStep) {
-    case stepsAddRol.generalInformation.id: {
-      const values = formReferences.generalInformation.current?.values;
-
-      if (!values) return addLinixRole;
-
-      newAddLinixRole.generalInformation = {
-        isValid: isCurrentFormValid,
-        values,
-      };
-
-      if (
-        JSON.stringify(values) !==
-        JSON.stringify(addLinixRole.generalInformation.values)
-      ) {
-        newAddLinixRole.generalInformation = {
-          isValid: false,
-          values: {
-            ...initialValuesAddRol.generalInformation,
-            roleName: values.roleName,
-            description: values.description,
-            aplication: values.aplication,
-          },
-        };
-      }
-
-      return newAddLinixRole;
-    }
-    case stepsAddRol.auxiliaryAccounts.id: {
-      const values = formReferences.ancillaryAccounts.current?.values;
-
-      if (!values) return addLinixRole;
-
-      newAddLinixRole.ancillaryAccounts = {
-        isValid: isCurrentFormValid,
-        values,
-      };
-
-      if (
-        JSON.stringify(values) !==
-        JSON.stringify(addLinixRole.generalInformation.values)
-      ) {
-        newAddLinixRole.ancillaryAccounts = {
-          isValid: false,
-          values: {
-            ...initialValuesAddRol.ancillaryAccounts,
-            officialSector: values.officialSector,
-            commercialSector: values.commercialSector,
-            solidaritySector: values.solidaritySector,
-          },
-        };
-      }
-
-      return newAddLinixRole;
-    }
-  }
-
-  const stepKey = Object.entries(stepsAddRol).find(
-    ([, config]) => config.id === currentStep
-  )?.[0];
-
-  if (!stepKey) return addLinixRole;
-
-  const values = formReferences[stepKey as keyof IFormAddRole]?.current?.values;
-
-  return (newAddLinixRole = {
-    ...newAddLinixRole,
-    [stepKey]: { isValid: isCurrentFormValid, values },
-  });
-};
-
-export interface IFormAddRole {
-  generalInformation: {
-    isValid: boolean;
-    values: IGeneralInformationForm;
-  };
-  ancillaryAccounts: {
-    isValid: boolean;
-    values: IAncillaryAccountsForm;
-  };
-}
+import { initialValuesAddRol } from "./config/initialValues";
 
 const steps = Object.values(stepsAddRol);
 
@@ -158,7 +51,7 @@ export function AddRol() {
   };
 
   const handleStepChange = (stepId: number) => {
-    const newCreditDestinationRequest = creditDestinationStepsRules(
+    const newCreditDestinationRequest = addRoleStepsRules(
       currentStep,
       dataAddRoleLinix,
       formReferences,
