+++ src/pages/privileges/outlets/roles/utils.ts
@@ -1,0 +1,82 @@
+import { stepsAddRol } from "./add-role/config/addRol.config";
+import { initialValuesAddRol } from "./add-role/config/initialValues";
+import { IFormAddRole, IFormAddRoleRef } from "./types";
+
+export const addRoleStepsRules = (
+  currentStep: number,
+  addLinixRole: IFormAddRole,
+  formReferences: IFormAddRoleRef,
+  isCurrentFormValid: boolean
+) => {
+  let newAddLinixRole = { ...addLinixRole };
+
+  switch (currentStep) {
+    case stepsAddRol.generalInformation.id: {
+      const values = formReferences.generalInformation.current?.values;
+
+      if (!values) return addLinixRole;
+
+      newAddLinixRole.generalInformation = {
+        isValid: isCurrentFormValid,
+        values,
+      };
+
+      if (
+        JSON.stringify(values) !==
+        JSON.stringify(addLinixRole.generalInformation.values)
+      ) {
+        newAddLinixRole.generalInformation = {
+          isValid: false,
+          values: {
+            ...initialValuesAddRol.generalInformation,
+            roleName: values.roleName,
+            description: values.description,
+            aplication: values.aplication,
+          },
+        };
+      }
+
+      return newAddLinixRole;
+    }
+    case stepsAddRol.auxiliaryAccounts.id: {
+      const values = formReferences.ancillaryAccounts.current?.values;
+
+      if (!values) return addLinixRole;
+
+      newAddLinixRole.ancillaryAccounts = {
+        isValid: isCurrentFormValid,
+        values,
+      };
+
+      if (
+        JSON.stringify(values) !==
+        JSON.stringify(addLinixRole.generalInformation.values)
+      ) {
+        newAddLinixRole.ancillaryAccounts = {
+          isValid: false,
+          values: {
+            ...initialValuesAddRol.ancillaryAccounts,
+            officialSector: values.officialSector,
+            commercialSector: values.commercialSector,
+            solidaritySector: values.solidaritySector,
+          },
+        };
+      }
+
+      return newAddLinixRole;
+    }
+  }
+
+  const stepKey = Object.entries(stepsAddRol).find(
+    ([, config]) => config.id === currentStep
+  )?.[0];
+
+  if (!stepKey) return addLinixRole;
+
+  const values = formReferences[stepKey as keyof IFormAddRole]?.current?.values;
+
+  return (newAddLinixRole = {
+    ...newAddLinixRole,
+    [stepKey]: { isValid: isCurrentFormValid, values },
+  });
+};
